run tests with and without MPI
